fix: repair non-UTF8 bytes in struct.css, split selection module, indent fixes
Fallow scan found invalid UTF-8 bytes (0x85, 0x97) in struct.css — re-saved as proper UTF-8.

Split src/webview/memory/selection.ts into data/query (selection.ts) and DOM-event (selectionClick.ts) layers to resolve fallow split_high_impact refactoring target.

Fixed 5-space indent of user-select: none across struct.css, memory-view.css, and base.css.
