+++ src/webview/hexViewer.ts
@@ -9,7 +9,8 @@ import { parsePasteText }                             from './pasteUtils';
 import { renderMemHeader, renderMemBody, applySel, scrollTo } from './memory/memoryView';
 import { setupMemoryDragSelection as setupMemoryDragSelectionController } from './memory/dragSelection';
 import { getByte } from './memory/memoryData';
-import { currentSelectionRange, selectedBytes, selectByteForContextMenu, selectByteFromClick } from './memory/selection';
+import { currentSelectionRange, selectedBytes } from './memory/selection';
+import { selectByteForContextMenu, selectByteFromClick } from './memory/selectionClick';
 import { renderInspector, renderBits, renderSegments, renderLabels, updateInspector, updateLabelFormSel } from './sidebar/sidebar';
 import { setupSidebarResize } from './sidebar/sidebarResize';
 import { renderStructPins, onSelectionChangeForStruct, resetStructViewState } from './sidebar/struct/index';
