+++ src/webview/styles/memory-view.css
@@ -8,7 +8,7 @@
     background: var(--toolbar-bg); border-bottom: 2px solid var(--border);
     font-family: var(--font-editor); font-weight: 700;
     color: var(--addr-active-fg);
-     user-select: none;
+    user-select: none;
     overflow: hidden; white-space: nowrap;
 }
 #mem-header .addr-cell { visibility: hidden; }
@@ -57,7 +57,7 @@
     font-family: var(--font-editor); font-size: var(--font-size);
     width: var(--cell-size); height: var(--cell-size); line-height: var(--cell-size);
     text-align: center; display: inline-block; border-radius: 2px;
-     user-select: none; cursor: pointer;
+    user-select: none; cursor: pointer;
 }
 /* 4-byte group gap: add left margin to the 5th, 9th, 13th cell (1-indexed) */
 .cell-group .data-cell:nth-child(4n+1):not(:first-child) { margin-left: calc(var(--cell-size) * 0.4); }
@@ -92,7 +92,7 @@
     font-family: var(--font-editor); font-size: var(--font-size);
     width: var(--text-cell-width); height: var(--cell-size); line-height: var(--cell-size);
     text-align: center; display: inline-block; border-radius: 2px;
-     user-select: none; cursor: pointer;
+    user-select: none; cursor: pointer;
 }
 .char-cell.cp { color: var(--fg); }
 .char-cell.cd { color: var(--non-graphic); cursor: default; }
@@ -153,7 +153,7 @@
     background: var(--gap-color);
     border-top: 1px dashed var(--gap-bdr); border-bottom: 1px dashed var(--gap-bdr);
     border-radius: 3px;
-     user-select: none; white-space: nowrap;
+    user-select: none; white-space: nowrap;
 }
 .gap-dots  { font-family: var(--font-editor); color: var(--gap-bdr); margin-right: 10px; letter-spacing: 4px; }
 .gap-range { font-family: var(--font-editor); font-size: var(--font-size); color: var(--addr-fg); }
@@ -164,7 +164,7 @@
     display: flex; align-items: center; gap: 10px;
     padding: 2px 10px 2px calc(var(--cell-pad) + 6px);
     border-left: 3px solid;
-     user-select: none;
+    user-select: none;
 }
 .seg-banner .sb-name { font-size: 11px; font-weight: 700; }
 .seg-banner .sb-meta { font-family: var(--font-editor); font-size: 10px; opacity: .65; }
